# systems/crawler/docs/openapi.yaml
openapi: 3.1.0
info:
  title: AACV System API
  version: 7.1.0
  description: 本契约描述认证、用户、审计、双源采集、MySQL 数据治理、规范目录、事务 Outbox、Neo4j 可重建投影、受限图查询、MySQL 聚合统计、异步导出和管理员运维接口；不包含服务器部署能力。
servers:
  - url: http://127.0.0.1:8080
    description: Windows 本地开发后端
...
paths:
...
  /api/v1/crawl/tasks:
    get:
      tags: [Crawl Tasks]
      summary: 分页查询采集任务
      operationId: listCrawlTasks
      security: [{sessionCookie: []}]
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Size'
      responses:
        '200':
          description: 任务分页
          content:
            application/json:
              schema: {$ref: '#/components/schemas/CrawlTaskPage'}
    post:
      tags: [Crawl Tasks]
      summary: 创建有界 OpenAlex 或 Crossref 采集任务
      operationId: createCrawlTask
      security: [{sessionCookie: []}]
      parameters:
        - $ref: '#/components/parameters/CsrfHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/CreateCrawlTaskRequest'}
      responses:
        '201':
          description: 任务已创建
          content:
            application/json:
              schema: {$ref: '#/components/schemas/CrawlTask'}
  /api/v1/crawl/tasks/{taskId}:
    get:
      tags: [Crawl Tasks]
      summary: 查询采集任务详情
      operationId: getCrawlTask
      security: [{sessionCookie: []}]
      parameters:
        - $ref: '#/components/parameters/TaskId'
      responses:
        '200':
          description: 任务详情
          content:
            application/json:
              schema: {$ref: '#/components/schemas/CrawlTask'}
    put:
      tags: [Crawl Tasks]
      summary: 更新尚未运行的采集任务
      operationId: updateCrawlTask
      security: [{sessionCookie: []}]
      parameters:
        - $ref: '#/components/parameters/TaskId'
        - $ref: '#/components/parameters/CsrfHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/UpdateCrawlTaskRequest'}
      responses:
        '200':
          description: 任务已更新
          content:
            application/json:
              schema: {$ref: '#/components/schemas/CrawlTask'}
        '409': {$ref: '#/components/responses/Conflict'}
  /api/v1/crawl/tasks/{taskId}/trigger:
    post:
      tags: [Crawl Tasks]
      summary: 手动触发一次采集运行
      operationId: triggerCrawlTask
      security: [{sessionCookie: []}]
      parameters:
        - $ref: '#/components/parameters/TaskId'
        - $ref: '#/components/parameters/CsrfHeader'
      responses:
        '202':
          description: 运行已创建
          content:
            application/json:
              schema: {$ref: '#/components/schemas/CrawlRun'}
        '409': {$ref: '#/components/responses/Conflict'}
  /api/v1/crawl/tasks/{taskId}/schedule:
    put:
      tags: [Crawl Tasks]
      summary: 配置每日固定查询范围复查计划
      operationId: configureDailySchedule
      security: [{sessionCookie: []}]
      parameters:
        - $ref: '#/components/parameters/TaskId'
        - $ref: '#/components/parameters/CsrfHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/DailyScheduleRequest'}
      responses:
        '200':
          description: 每日计划已配置
          content:
            application/json:
              schema: {$ref: '#/components/schemas/CrawlSchedule'}
  /api/v1/crawl/runs/{runId}:
    get:
      tags: [Crawl Runs]
      summary: 查询运行详情和检查点统计
      operationId: getCrawlRun
      security: [{sessionCookie: []}]
      parameters:
        - $ref: '#/components/parameters/RunId'
      responses:
        '200':
          description: 运行详情
          content:
            application/json:
              schema: {$ref: '#/components/schemas/CrawlRun'}
  /api/v1/crawl/runs/{runId}/failures:
    get:
      tags: [Crawl Runs]
      summary: 分页查询运行失败记录
      operationId: listCrawlFailures
      security: [{sessionCookie: []}]
      parameters:
        - $ref: '#/components/parameters/RunId'
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Size'
      responses:
        '200':
          description: 失败记录分页
          content:
            application/json:
              schema: {$ref: '#/components/schemas/CrawlFailurePage'}
  /api/v1/crawl/runs/{runId}/pause:
    post:
      tags: [Crawl Runs]
      summary: 请求在当前 Chunk 提交后暂停
      operationId: pauseCrawlRun
      security: [{sessionCookie: []}]
      parameters:
        - $ref: '#/components/parameters/RunId'
        - $ref: '#/components/parameters/CsrfHeader'
      responses:
        '202': {description: 暂停意图已记录}
  /api/v1/crawl/runs/{runId}/resume:
    post:
      tags: [Crawl Runs]
      summary: 从最近成功检查点恢复运行
      operationId: resumeCrawlRun
      security: [{sessionCookie: []}]
      parameters:
        - $ref: '#/components/parameters/RunId'
        - $ref: '#/components/parameters/CsrfHeader'
      responses:
        '202': {description: 恢复已请求}
  /api/v1/crawl/runs/{runId}/cancel:
    post:
      tags: [Crawl Runs]
      summary: 请求在当前 Chunk 提交后取消
      operationId: cancelCrawlRun
      security: [{sessionCookie: []}]
      parameters:
        - $ref: '#/components/parameters/RunId'
        - $ref: '#/components/parameters/CsrfHeader'
      responses:
        '202': {description: 取消意图已记录}
  /api/v1/crawl/runs/{runId}/retry-failures:
    post:
      tags: [Crawl Runs]
      summary: 有限重试失败记录并保留原运行证据
      operationId: retryCrawlFailures
      security: [{sessionCookie: []}]
      parameters:
        - $ref: '#/components/parameters/RunId'
        - $ref: '#/components/parameters/CsrfHeader'
      responses:
        '202': {description: 重试运行已创建}
...
  /api/v1/duplicate-candidates:
    get:
      tags: [Governance]
      summary: 分页查询重复候选
      operationId: listDuplicateCandidates
      security: [{sessionCookie: []}]
      parameters:
        - {name: entityType, in: query, schema: {type: string, enum: [ACHIEVEMENT, AUTHOR, ORGANIZATION, VENUE]}}
        - {name: status, in: query, schema: {type: string, enum: [PENDING, ACCEPTED, REJECTED]}}
        - {name: sourceId, in: query, schema: {type: integer, format: int64, minimum: 1}}
        - {name: ruleVersion, in: query, schema: {type: integer, minimum: 1}}
        - {name: createdFrom, in: query, schema: {type: string, format: date-time}}
        - {name: createdTo, in: query, schema: {type: string, format: date-time}}
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Size'
      responses:
        '200':
          description: 仅含有限匹配证据的候选分页
          content:
            application/json:
              schema: {$ref: '#/components/schemas/DuplicateCandidatePage'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '403': {$ref: '#/components/responses/Forbidden'}
  /api/v1/duplicate-candidates/{candidateId}:
    get:
      tags: [Governance]
      summary: 查询重复候选详情
      operationId: getDuplicateCandidate
      security: [{sessionCookie: []}]
      parameters:
        - {name: candidateId, in: path, required: true, schema: {type: integer, format: int64, minimum: 1}}
      responses:
        '200':
          description: 候选详情
          content:
            application/json:
              schema: {$ref: '#/components/schemas/DuplicateCandidate'}
        '404': {$ref: '#/components/responses/NotFound'}
  /api/v1/duplicate-candidates/{candidateId}/comparison:
    get:
      tags: [Governance]
      summary: 对照候选两侧当前字段与明确版本关系
      operationId: compareDuplicateCandidate
      x-required-permission: GOVERNANCE_READ
      security: [{sessionCookie: []}]
      parameters:
        - {name: candidateId, in: path, required: true, schema: {type: integer, format: int64, minimum: 1}}
      responses:
        '200':
          description: 两侧实体的有限字段；未知值保留null，标题包含有效人工覆盖
          content:
            application/json:
              schema: {$ref: '#/components/schemas/CandidateComparison'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '404': {$ref: '#/components/responses/NotFound'}
  /api/v1/duplicate-candidates/{candidateId}/accept:
    post:
      tags: [Governance]
      summary: 接受候选并建立逻辑规范关联
      operationId: acceptDuplicateCandidate
      security: [{sessionCookie: []}]
      parameters:
        - {name: candidateId, in: path, required: true, schema: {type: integer, format: int64, minimum: 1}}
        - $ref: '#/components/parameters/CsrfHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/CandidateAcceptRequest'}
      responses:
        '200': {description: 接受决定, content: {application/json: {schema: {$ref: '#/components/schemas/MergeDecision'}}}}
        '409': {$ref: '#/components/responses/Conflict'}
  /api/v1/duplicate-candidates/{candidateId}/reject:
    post:
      tags: [Governance]
      summary: 拒绝候选并抑制同规则重复生成
      operationId: rejectDuplicateCandidate
      security: [{sessionCookie: []}]
      parameters:
        - {name: candidateId, in: path, required: true, schema: {type: integer, format: int64, minimum: 1}}
        - $ref: '#/components/parameters/CsrfHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/VersionReasonRequest'}
      responses:
        '200': {description: 拒绝决定, content: {application/json: {schema: {$ref: '#/components/schemas/MergeDecision'}}}}
        '409': {$ref: '#/components/responses/Conflict'}
...
  /api/v1/exports:
    post:
      tags: [Exports]
      summary: 创建当前筛选结果的异步导出
      operationId: createExport
      x-required-permission: EXPORT_CREATE
      security: [{sessionCookie: []}]
      parameters: [{$ref: '#/components/parameters/CsrfHeader'}]
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/ExportCreateRequest'}
      responses:
        '202': {description: 已创建导出任务, content: {application/json: {schema: {$ref: '#/components/schemas/ExportTask'}}}}
        '400': {$ref: '#/components/responses/BadRequest'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '409': {$ref: '#/components/responses/Conflict'}
  /api/v1/exports/{exportId}:
    get:
      tags: [Exports]
      summary: 查询本人创建的导出任务状态，管理员可查询全部任务
      operationId: getExport
      x-required-permission: EXPORT_READ
      security: [{sessionCookie: []}]
      parameters:
        - {name: exportId, in: path, required: true, schema: {type: string, format: uuid}}
      responses:
        '200': {description: 导出任务状态, content: {application/json: {schema: {$ref: '#/components/schemas/ExportTask'}}}}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '404': {$ref: '#/components/responses/NotFound'}
  /api/v1/exports/{exportId}/download:
    get:
      tags: [Exports]
      summary: 下载本人已完成且未过期的导出文件，管理员可下载全部任务
      operationId: downloadExport
      x-required-permission: EXPORT_READ
      security: [{sessionCookie: []}]
      parameters:
        - {name: exportId, in: path, required: true, schema: {type: string, format: uuid}}
        - {name: token, in: query, required: true, schema: {type: string, minLength: 32, maxLength: 128}}
      responses:
        '200':
          description: 导出文件
          content:
            text/csv: {schema: {type: string, format: binary}}
            application/json: {schema: {type: string, format: binary}}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '404': {$ref: '#/components/responses/NotFound'}
        '410': {$ref: '#/components/responses/Gone'}
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: SESSION
  parameters:
    CsrfHeader:
      name: X-CSRF-TOKEN
      in: header
      required: true
      schema:
        type: string
        minLength: 1
...
    TaskId:
      name: taskId
      in: path
      required: true
      schema: {type: integer, format: int64, minimum: 1}
    RunId:
      name: runId
      in: path
      required: true
      schema: {type: integer, format: int64, minimum: 1}
...
    Page:
      name: page
      in: query
      schema:
        type: integer
        minimum: 0
        default: 0
    Size:
      name: size
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
...
  responses:
...
    BadRequest:
      description: 请求校验失败
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    Unauthorized:
      description: 未登录、凭据错误或会话失效
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    Forbidden:
      description: 权限不足或 CSRF Token 无效
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    Conflict:
      description: 资源状态、并发限制、用户名或乐观锁版本冲突
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    NotFound:
      description: 请求的资源不存在
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    Gone:
      description: 请求的导出文件已经过期
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
...
  schemas:
...
    CrawlTaskParameters:
      type: object
      additionalProperties: false
      required: [maxPages, maxRecords]
      properties:
        publicationDateFrom: {type: [string, 'null'], format: date}
        publicationDateTo: {type: [string, 'null'], format: date}
        keyword: {type: [string, 'null'], maxLength: 200}
        authorIds: {type: array, maxItems: 50, items: {type: string, maxLength: 128}}
        institutionIds: {type: array, maxItems: 50, items: {type: string, maxLength: 128}}
        dois: {type: array, maxItems: 50, items: {type: string, maxLength: 255}}
        orcids: {type: array, maxItems: 50, items: {type: string, maxLength: 64}}
        rorIds: {type: array, maxItems: 50, items: {type: string, maxLength: 128}}
        updatedFrom: {type: [string, 'null'], format: date-time}
        updatedUntil: {type: [string, 'null'], format: date-time}
        maxPages: {type: integer, minimum: 1, maximum: 5}
        maxRecords: {type: integer, minimum: 1, maximum: 500}
    CreateCrawlTaskRequest:
      type: object
      additionalProperties: false
      required: [sourceId, name, parameters]
      properties:
        sourceId: {type: integer, format: int64, minimum: 1}
        name: {type: string, minLength: 1, maxLength: 128}
        parameters: {$ref: '#/components/schemas/CrawlTaskParameters'}
    UpdateCrawlTaskRequest:
      type: object
      additionalProperties: false
      required: [name, parameters, version]
      properties:
        name: {type: string, minLength: 1, maxLength: 128}
        parameters: {$ref: '#/components/schemas/CrawlTaskParameters'}
        version: {type: integer, format: int64, minimum: 0}
    CrawlTask:
      type: object
      required: [id, sourceId, name, parameterVersion, parameters, enabled, version, createdAt, updatedAt]
      properties:
        id: {type: integer, format: int64}
        sourceId: {type: integer, format: int64}
        name: {type: string}
        parameterVersion: {type: integer, enum: [1, 2], description: 版本 1 保持 OpenAlex 既有语义，版本 2 承载 Crossref 扩展过滤条件}
        parameters: {$ref: '#/components/schemas/CrawlTaskParameters'}
        enabled: {type: boolean}
        version: {type: integer, format: int64}
        createdAt: {type: string, format: date-time}
        updatedAt: {type: string, format: date-time}
    CrawlTaskPage:
      allOf:
        - $ref: '#/components/schemas/PageMetadata'
        - type: object
          properties:
            items: {type: array, items: {$ref: '#/components/schemas/CrawlTask'}}
    DailyScheduleRequest:
      type: object
      additionalProperties: false
      required: [localTime, timeZone]
      properties:
        localTime: {type: string, pattern: '^(?:[01]\\d|2[0-3]):[0-5]\\d$'}
        timeZone: {type: string, maxLength: 64}
        version: {type: integer, format: int64, minimum: 0}
    CrawlSchedule:
      type: object
      required: [taskId, localTime, timeZone, nextFireAt, version, incrementalMode]
      properties:
        taskId: {type: integer, format: int64}
        localTime: {type: string}
        timeZone: {type: string}
        nextFireAt: {type: string, format: date-time}
        version: {type: integer, format: int64}
        incrementalMode:
          type: string
          enum: [FIXED_SCOPE_REFRESH, ROLLING_PUBLICATION_DATE_WINDOW, CLOSED_INDEX_DATE_WINDOW]
          description: 每日计划当前重复固定查询范围，返回FIXED_SCOPE_REFRESH；保留历史模式名供旧数据兼容，不承诺自动移动日期窗口。
    CrawlRun:
      type: object
      required: [id, taskId, runNumber, triggerType, status, readCount, parsedCount, createdCount, updatedCount, duplicateCount, failureCount, requestCount]
      properties:
        id: {type: integer, format: int64}
        taskId: {type: integer, format: int64}
        runNumber: {type: string}
        triggerType: {type: string, enum: [MANUAL, SCHEDULED, RESUME, RETRY_FAILURES]}
        parentRunId: {type: [integer, 'null'], format: int64}
        status: {type: string, enum: [PENDING, RUNNING, PAUSING, PAUSED, SUCCEEDED, PARTIAL_SUCCESS, FAILED, CANCELLING, CANCELLED]}
        batchJobExecutionId: {type: [integer, 'null'], format: int64}
        readCount: {type: integer, format: int64}
        parsedCount: {type: integer, format: int64}
        createdCount: {type: integer, format: int64}
        updatedCount: {type: integer, format: int64}
        duplicateCount: {type: integer, format: int64}
        failureCount: {type: integer, format: int64}
        requestCount: {type: integer, format: int64}
        checkpoint: {type: [string, 'null']}
        startedAt: {type: [string, 'null'], format: date-time}
        finishedAt: {type: [string, 'null'], format: date-time}
        completionReason:
          type: [string, 'null']
          enum: [SOURCE_EXHAUSTED, PAGE_LIMIT, RECORD_LIMIT, RETRY_BATCH_COMPLETED, QUOTA_EXHAUSTED, QUOTA_RETRY_LIMIT, USER_PAUSED, USER_CANCELLED, BATCH_FAILED, null]
          description: 来源耗尽仅指当前查询范围；达到页数或记录上限返回PARTIAL_SUCCESS，不代表范围完整。
        deferredUntil:
          type: [string, 'null']
          format: date-time
          description: 每日额度恢复时间；只有PAUSED且QUOTA_EXHAUSTED的运行由服务端自动恢复。
        quotaDeferrals: {type: integer, minimum: 0, maximum: 3}
...
    CrawlFailurePage:
      allOf:
        - $ref: '#/components/schemas/PageMetadata'
        - type: object
          properties:
            items: {type: array, items: {$ref: '#/components/schemas/CrawlFailure'}}
...
    PageMetadata:
      type: object
      required: [items, page, size, totalElements, totalPages]
      properties:
        page: {type: integer}
        size: {type: integer}
        totalElements: {type: integer, format: int64}
        totalPages: {type: integer}
...
    ExportCreateRequest:
      type: object
      additionalProperties: false
      required: [format, filters]
      properties:
        format: {$ref: '#/components/schemas/ExportFormat'}
        filters: {$ref: '#/components/schemas/ExportFilter'}
    ExportTask:
      type: object
      required: [id, format, status, requestedBy, requestedCount, exportedCount, downloadAvailable, downloadToken, createdAt, startedAt, completedAt, expiresAt, errorCode, errorMessage]
      properties:
        id: {type: string, format: uuid}
        format: {$ref: '#/components/schemas/ExportFormat'}
        status: {$ref: '#/components/schemas/ExportStatus'}
        requestedBy: {type: integer, format: int64, minimum: 1}
        requestedCount: {type: integer, format: int64, minimum: 0, maximum: 10000}
        exportedCount: {type: integer, format: int64, minimum: 0, maximum: 10000}
        downloadAvailable: {type: boolean}
        downloadToken: {type: [string, 'null'], minLength: 32, maxLength: 128, description: 仅成功且未过期的任务返回服务端生成的不可预测令牌}
        createdAt: {type: string, format: date-time}
        startedAt: {type: [string, 'null'], format: date-time}
        completedAt: {type: [string, 'null'], format: date-time}
        expiresAt: {type: [string, 'null'], format: date-time}
        errorCode: {type: [string, 'null']}
        errorMessage: {type: [string, 'null'], description: 有限安全摘要，不包含文件路径、SQL或敏感配置}
...
    Problem:
      type: object
      required: [type, title, status, detail, instance, errorCode, traceId]
      properties:
        type:
          type: string
          format: uri-reference
        title:
          type: string
        status:
          type: integer
        detail:
          type: string
        instance:
          type: string
          format: uri-reference
        errorCode:
          type: string
        traceId:
          type: string
        fieldErrors:
          type: array
          items:
            type: object
            required: [field, message]
            properties:
              field:
                type: string
              message:
                type: string
...
    CandidateComparison:
      type: object
      required: [candidateId, candidateVersion, entityType, leftEntityId, rightEntityId, left, right, explicitVersionRelation]
      properties:
        candidateId: {type: integer, format: int64}
        candidateVersion: {type: integer, format: int64}
        entityType: {type: string, enum: [ACHIEVEMENT, AUTHOR, ORGANIZATION, VENUE]}
        leftEntityId: {type: integer, format: int64}
        rightEntityId: {type: integer, format: int64}
        left: {type: object, additionalProperties: true}
        right: {type: object, additionalProperties: true}
        explicitVersionRelation:
          type: boolean
          description: 来源快照明确连接两侧DOI的版本关系；为true时接受合并返回409，可拒绝候选
    DuplicateCandidate:
      type: object
      required: [id, entityType, leftEntityId, rightEntityId, matchBasis, evidence, status, ruleVersion, version, createdAt]
      properties:
        id: {type: integer, format: int64}
        entityType: {type: string, enum: [ACHIEVEMENT, AUTHOR, ORGANIZATION, VENUE]}
        leftEntityId: {type: integer, format: int64}
        rightEntityId: {type: integer, format: int64}
        matchBasis: {type: string, enum: [FINGERPRINT, TEXT_NAME, SOURCE_POSITION]}
        evidence: {type: object, additionalProperties: true, description: 有限匹配证据，不含原始Payload}
        status: {type: string, enum: [PENDING, ACCEPTED, REJECTED]}
        sourceId: {type: [integer, 'null'], format: int64}
        ruleVersion: {type: integer, minimum: 1}
        version: {type: integer, format: int64, minimum: 0}
        createdAt: {type: string, format: date-time}
        updatedAt: {type: string, format: date-time}
    DuplicateCandidatePage:
      allOf:
        - $ref: '#/components/schemas/PageMetadata'
        - type: object
          properties:
            items: {type: array, items: {$ref: '#/components/schemas/DuplicateCandidate'}}
    CandidateAcceptRequest:
      type: object
      required: [canonicalEntityId, reason, version]
      properties:
        canonicalEntityId: {type: integer, format: int64, minimum: 1}
        reason: {type: string, minLength: 1, maxLength: 1000}
        version: {type: integer, format: int64, minimum: 0}
    VersionReasonRequest:
      type: object
      required: [reason, version]
      properties:
        reason: {type: string, minLength: 1, maxLength: 1000}
        version: {type: integer, format: int64, minimum: 0}
    MergeDecision:
      type: object
      required: [id, candidateId, decision, revisionId, actorUserId, reason, version, decidedAt]
      properties:
        id: {type: integer, format: int64}
        candidateId: {type: integer, format: int64}
        decision: {type: string, enum: [ACCEPT, REJECT, REVERT]}
        canonicalEntityId: {type: [integer, 'null'], format: int64}
        revisionId: {type: integer, format: int64}
        actorUserId: {type: integer, format: int64}
        reason: {type: string}
        version: {type: integer, format: int64}
        decidedAt: {type: string, format: date-time}
...